+++ contracts/utils/VestingLib.sol
@@ -26,7 +26,7 @@ import {SafeCast} from "../vendor/openzeppelin/v4.9.0/utils/math/SafeCast.sol";
  */
 
 abstract contract VestingLib {
-    uint32 public index;
+    uint256 public index;
 
     struct VestingPosition {
         address beneficiary;
@@ -38,10 +38,10 @@ abstract contract VestingLib {
 
     event NewVestingPosition (
         VestingPosition position,
-        uint32 index
+        uint256 index
     );
 
-    mapping(uint32 => VestingPosition) internal vestingPositions;
+    mapping(uint256 => VestingPosition) internal vestingPositions;
 
     error VestingLib__StartTimestampMustNotBeInThePast();
     error VestingLib__InvalidDuration();
@@ -56,13 +56,13 @@ abstract contract VestingLib {
      *
      * @return uint256 The amount that has vested but hasn't been released yet
      */
-    function releasableAmount(uint32 vestingIndex) public view returns (uint256) {
+    function releasableAmount(uint256 vestingIndex) public view returns (uint256) {
         VestingPosition memory vestingPosition = vestingPositions[vestingIndex];
         return vestedAmount(vestingIndex) - vestingPosition.releasedAmount;
     }
 
     function vestedAmount(
-        uint32 vestingIndex
+        uint256 vestingIndex
     ) internal view returns (uint256 vestedAmount_) {
         VestingPosition memory vestingPosition_ = vestingPositions[
             vestingIndex
@@ -116,7 +116,7 @@ abstract contract VestingLib {
 
     // @dev This is a function which should be called when user claims some amount of releasable tokens
     function updateReleasedAmount(
-        uint32 vestingIndex,
+        uint256 vestingIndex,
         uint256 releaseAmount
     ) internal {
         if (vestingIndex >= index) revert VestingLib__InvalidIndex();
